Update and clean validation error handlers tests
