+++ .badges/tests.json
@@ -1,7 +1,7 @@
 {
   "schemaVersion": 1,
   "label": "tests",
-  "message": "1432",
+  "message": "1434",
   "color": "blue",
   "cacheSeconds": 300
 }
